@@ -6,23 +6,23 @@
     "start": "npx next start"
   },
   "dependencies": {
-    "bootstrap": "^5.1.3",
-    "next": "13.0.5",
-    "ramda": "^0.28.0",
+    "bootstrap": "^5.3.0",
+    "next": "13.4.4",
+    "ramda": "^0.29.0",
     "react": "18.2.0",
-    "react-bootstrap": "^2.2.1",
+    "react-bootstrap": "^2.7.4",
     "react-dom": "18.2.0",
-    "react-hook-form": "^7.32.0"
+    "react-hook-form": "^7.44.3"
   },
   "devDependencies": {
     "@fullhuman/postcss-purgecss": "^4.1.3",
     "@types/lodash": "^4.14.181",
-    "@types/node": "^17.0.21",
+    "@types/node": "^18.11.19",
     "@types/ramda": "^0.28.7",
-    "@types/react": "^17.0.40",
+    "@types/react": "^18.2.8",
     "postcss-flexbugs-fixes": "^5.0.2",
-    "postcss-preset-env": "^7.4.3",
-    "sass": "^1.49.10",
-    "typescript": "^4.6.2"
+    "postcss-preset-env": "^8.4.2",
+    "sass": "^1.62.1",
+    "typescript": "^5.1.3"
   }
 }
