@@ -1,5 +1,4 @@
-import { useRouter } from "next/router";
-import { ParsedUrlQueryInput } from "node:querystring";
+import { ReadonlyURLSearchParams, usePathname, useRouter, useSearchParams } from "next/navigation";
 import { FormType } from "@type/formType";
 import { OrderByEnum } from "@type/sortType";
 
@@ -19,34 +18,48 @@ type ACTIONTYPE = {
     } & Partial<FormType>;
 };
 
-const queryReducer = (state: ParsedUrlQueryInput, action: ACTIONTYPE): ParsedUrlQueryInput => {
+const queryReducer = (searchParams: ReadonlyURLSearchParams, action: ACTIONTYPE): URLSearchParams => {
     const { type, payload: { orderAsc, orderBy, page } = {} } = action;
-
+    const params = new URLSearchParams(searchParams);
     switch (type) {
         case ACTIONS_QUERY.CHANGE_PAGE:
-            return { ...state, page };
+            params.set('page', page!.toString());
+            return params;
         case ACTIONS_QUERY.RESET:
-            return {};
+            params.delete('country');
+            params.delete('greaterThan');
+            params.delete('lessThan');
+            return params;
         case ACTIONS_QUERY.SORT:
-            return { ...state, orderBy, orderAsc, page: 1 };
+            if(orderBy) params.set('orderBy', orderBy as string);
+            if(orderAsc?.toString()) params.set('orderAsc', orderAsc.toString());
+            params.set('page', '1');
+            return params;
         case ACTIONS_QUERY.SUBMIT:
-            const dataToURL = Object.fromEntries(Object.entries(action.payload).filter(([_, v]) => v != ""));
-            return { ...state, ...dataToURL, page: 1 };
+            for (const [key, value] of Object.entries(action.payload!)) {
+                if (value != "") {
+                    params.set(key, value as string);
+                } else {
+                    params.delete(key);
+                }
+            }
+            return params;
         default:
-            return state;
+            return params;
     }
 };
 
 const useQueryAction = () => {
     const router = useRouter();
-    const { pathname, query } = router;
+    const pathname = usePathname();
+    const searchParams = useSearchParams();
 
-    const dispatchQuery = (action: ACTIONTYPE) => router.push({ pathname, query: queryReducer(query, action) }, undefined, { shallow: true });
+    const dispatchQuery = (action: ACTIONTYPE) => router.replace(`${pathname}?${queryReducer(searchParams, action).toString()}`);
 
     const sortLinkProp = (orderBy: `${OrderByEnum}`) => ({
         href: {
             pathname,
-            query: queryReducer(query, { type: ACTIONS_QUERY.SORT, payload: { orderBy, orderAsc: query.orderAsc !== "true" } }),
+            query: Object.fromEntries(queryReducer(searchParams, { type: ACTIONS_QUERY.SORT, payload: { orderBy, orderAsc: searchParams.get('orderAsc') !== "true" } })),
         },
         shallow: true,
     });
